More Fixes.
The changes that have been made allow the code to represent which choices the user got wrong when creating their bracket. It displays the correct choices in green, and the incorrect choices in red. Also, The code has been modified to allow a user to view the scoreboard before logging in.